# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Movena replacement core contracts",
    "version": "2.0.0-alpha.1",
    "description": "Typed initial recovery slice. Other compatibility endpoints remain inventoried in baseline.json."
  },
  "paths": {
...
    "/api/v2/therapist/appointments/{appointment_id}/session-notes": {
      "get": {
        "operationId": "getStaffVisitNotes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClinicalNoteDetail"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "operationId": "createVisitNote",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClinicalNoteDetail"
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClinicalNoteCreate"
              }
            }
          }
        }
      }
    },
...
    "/api/v2/therapist/patients/{patient_id}/reports": {
      "post": {
        "operationId": "createProgressReport",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressReportCreateResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "period_start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "period_end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "share_with_patient",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ]
      }
    },
...
    "/api/v2/therapist/appointments": {
      "get": {
        "operationId": "getStaffAppointments",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppointmentSummary"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      }
    },
...
    "/api/v2/therapist/availability": {
      "get": {
        "operationId": "getAvailability",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AvailabilityDetail"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      },
      "post": {
        "operationId": "addAvailability",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityDetail"
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvailabilityCreate"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ClinicalNoteCreate": {
        "properties": {
          "summary": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Summary",
            "type": "string"
          },
          "recommendations": {
            "anyOf": [
              {
                "maxLength": 10000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Recommendations"
          },
          "patient_visible": {
            "default": false,
            "title": "Patient Visible",
            "type": "boolean"
          }
        },
        "required": [
          "summary"
        ],
        "title": "ClinicalNoteCreate",
        "type": "object"
      },
      "ClinicalNoteDetail": {
        "properties": {
          "summary": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Summary",
            "type": "string"
          },
          "recommendations": {
            "anyOf": [
              {
                "maxLength": 10000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Recommendations"
          },
          "patient_visible": {
            "default": false,
            "title": "Patient Visible",
            "type": "boolean"
          },
          "note_id": {
            "title": "Note Id",
            "type": "string"
          },
          "appointment_id": {
            "title": "Appointment Id",
            "type": "string"
          },
          "therapist_user_id": {
            "title": "Therapist User Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "author_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Author Name"
          }
        },
        "required": [
          "summary",
          "note_id",
          "appointment_id",
          "therapist_user_id",
          "created_at",
          "updated_at"
        ],
        "title": "ClinicalNoteDetail",
        "type": "object"
      },
...
      "ProgressReportCreateResponse": {
        "properties": {
          "progress_report_id": {
            "title": "Progress Report Id",
            "type": "string"
          },
          "download_url": {
            "title": "Download Url",
            "type": "string"
          },
          "shared_with_patient": {
            "title": "Shared With Patient",
            "type": "boolean"
          }
        },
        "required": [
          "progress_report_id",
          "download_url",
          "shared_with_patient"
        ],
        "title": "ProgressReportCreateResponse",
        "type": "object"
      },
...
      "AppointmentSummary": {
        "properties": {
          "appointment_id": {
            "title": "Appointment Id",
            "type": "string"
          },
          "patient_id": {
            "title": "Patient Id",
            "type": "string"
          },
          "therapist_user_id": {
            "title": "Therapist User Id",
            "type": "string"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          },
          "ends_at": {
            "format": "date-time",
            "title": "Ends At",
            "type": "string"
          },
          "status": {
            "enum": [
              "scheduled",
              "confirmed",
              "completed",
              "cancelled",
              "no_show"
            ],
            "title": "Status",
            "type": "string"
          },
          "delivery_mode": {
            "enum": [
              "video",
              "in_person"
            ],
            "title": "Delivery Mode",
            "type": "string"
          },
          "payment_status": {
            "title": "Payment Status",
            "type": "string"
          },
          "can_join": {
            "default": false,
            "title": "Can Join",
            "type": "boolean"
          }
        },
        "required": [
          "appointment_id",
          "patient_id",
          "therapist_user_id",
          "starts_at",
          "ends_at",
          "status",
          "delivery_mode",
          "payment_status"
        ],
        "title": "AppointmentSummary",
        "type": "object"
      },
...
      "AvailabilityCreate": {
        "properties": {
          "weekday": {
            "maximum": 6,
            "minimum": 0,
            "title": "Weekday",
            "type": "integer"
          },
          "start_minute": {
            "maximum": 1439,
            "minimum": 0,
            "title": "Start Minute",
            "type": "integer"
          },
          "end_minute": {
            "maximum": 1440,
            "minimum": 1,
            "title": "End Minute",
            "type": "integer"
          },
          "timezone_name": {
            "default": "Africa/Cairo",
            "maxLength": 64,
            "title": "Timezone Name",
            "type": "string"
          }
        },
        "required": [
          "weekday",
          "start_minute",
          "end_minute"
        ],
        "title": "AvailabilityCreate",
        "type": "object"
      },
      "AvailabilityDetail": {
        "properties": {
          "weekday": {
            "maximum": 6,
            "minimum": 0,
            "title": "Weekday",
            "type": "integer"
          },
          "start_minute": {
            "maximum": 1439,
            "minimum": 0,
            "title": "Start Minute",
            "type": "integer"
          },
          "end_minute": {
            "maximum": 1440,
            "minimum": 1,
            "title": "End Minute",
            "type": "integer"
          },
          "timezone_name": {
            "default": "Africa/Cairo",
            "maxLength": 64,
            "title": "Timezone Name",
            "type": "string"
          },
          "availability_id": {
            "title": "Availability Id",
            "type": "string"
          },
          "therapist_user_id": {
            "title": "Therapist User Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          }
        },
        "required": [
          "weekday",
          "start_minute",
          "end_minute",
          "availability_id",
          "therapist_user_id",
          "is_active"
        ],
        "title": "AvailabilityDetail",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}